+++ lib/src/models/album.dart
@@ -4,11 +4,14 @@
 part of '_models.dart';
 
 /// Json representation of an album
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class Album extends AlbumSimple {
   Album();
 
   factory Album.fromJson(Map<String, dynamic> json) => _$AlbumFromJson(json);
+
+  @override
+  Map<String, dynamic> toJson() => _$AlbumToJson(this);
 
   /// The copyright statements of the album.
   List<Copyright>? copyrights;
@@ -34,17 +37,23 @@ class Album extends AlbumSimple {
   int? popularity;
 }
 
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class AlbumSimple extends Object {
   AlbumSimple();
 
   factory AlbumSimple.fromJson(Map<String, dynamic> json) =>
       _$AlbumSimpleFromJson(json);
 
+  Map<String, dynamic> toJson() => _$AlbumSimpleToJson(this);
+
   /// Helper function that unwraps the items from the paging object.
   static Iterable<TrackSimple> _extractTracksFromPage(dynamic json) {
     if (json == null) {
       return [];
+    }
+
+    if (json is List) {
+      return json.map((trackJson) => TrackSimple.fromJson(trackJson));
     }
 
     var items = Paging.fromJson(json).itemsNative;