+++ lib/src/models/episode.dart
@@ -4,7 +4,7 @@
 part of '_models.dart';
 
 /// Json representation of an episode
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class Episode extends Object {
   Episode();
 
@@ -75,10 +75,12 @@ class Episode extends Object {
 
   factory Episode.fromJson(Map<String, dynamic> json) =>
       _$EpisodeFromJson(json);
+
+  Map<String, dynamic> toJson() => _$EpisodeToJson(this);
 }
 
 /// Json representation of an episode with information about its show
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class EpisodeFull extends Episode {
   EpisodeFull();
 
@@ -86,4 +88,7 @@ class EpisodeFull extends Episode {
 
   factory EpisodeFull.fromJson(Map<String, dynamic> json) =>
       _$EpisodeFullFromJson(json);
+
+  @override
+  Map<String, dynamic> toJson() => _$EpisodeFullToJson(this);
 }