+++ lib/src/models/followers.dart
@@ -4,12 +4,14 @@
 part of '_models.dart';
 
 /// Json representation of followers
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class Followers extends Object {
   Followers();
 
   factory Followers.fromJson(Map<String, dynamic> json) =>
       _$FollowersFromJson(json);
+
+  Map<String, dynamic> toJson() => _$FollowersToJson(this);
 
   /// A link to the Web API endpoint providing full details of the followers;
   /// null if not available.