+++ lib/src/models/queue.dart
@@ -1,11 +1,13 @@
 part of '_models.dart';
 
 /// Json representation of the playback queue
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class Queue extends Object {
   Queue();
 
   factory Queue.fromJson(Map<String, dynamic> json) => _$QueueFromJson(json);
+
+  Map<String, dynamic> toJson() => _$QueueToJson(this);
 
   /// The currently playing track in the queue
   @JsonKey(name: 'currently_playing')