+++ lib/src/models/user.dart
@@ -4,11 +4,14 @@
 part of '_models.dart';
 
 /// Json representation of a user
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class User extends Object implements UserPublic {
   User();
 
   factory User.fromJson(Map<String, dynamic> json) => _$UserFromJson(json);
+
+  @override
+  Map<String, dynamic> toJson() => _$UserToJson(this);
 
   /// The user's date-of-birth.
   ///
@@ -73,12 +76,14 @@ class User extends Object implements UserPublic {
 }
 
 /// Json representation of a publicly displayed user
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class UserPublic extends Object {
   UserPublic();
 
   factory UserPublic.fromJson(Map<String, dynamic> json) =>
       _$UserPublicFromJson(json);
+
+  Map<String, dynamic> toJson() => _$UserPublicToJson(this);
 
   /// The name displayed on the user's profile. null if not available.
   @JsonKey(name: 'display_name')
@@ -108,12 +113,14 @@ class UserPublic extends Object {
 }
 
 /// Json representation of the playback history
-@JsonSerializable(createToJson: false)
+@JsonSerializable()
 class PlayHistory extends Object {
   PlayHistory();
 
   factory PlayHistory.fromJson(Map<String, dynamic> json) =>
       _$PlayHistoryFromJson(json);
+
+  Map<String, dynamic> toJson() => _$PlayHistoryToJson(this);
 
   /// The track the user listened to.
   TrackSimple? track;