+++ test/spotify_test.dart
@@ -1,4 +1,6 @@
 import 'dart:async';
+import 'package:http_interceptor/http_interceptor.dart';
+
 import 'spotify_mock.dart';
 import 'package:test/test.dart';
 import 'package:spotify/spotify.dart';
@@ -209,6 +211,16 @@ Future main() async {
       expect(result['1'], isTrue);
       expect(result['2'], isFalse);
       expect(result['3'], isTrue);
+    });
+
+    test('get playlist tracks', () async {
+      var tracks = await spotify.playlists
+          .getTracksByPlaylistId('1XIAxOGAEK2h4ravpNTmYF')
+          .all();
+
+      expect(tracks, hasLength(2));
+      expect(tracks.elementAt(0).id, 'track-1');
+      expect(tracks.elementAt(1).id, 'track-3');
     });
   });
 
@@ -488,8 +500,8 @@ Future main() async {
       spotify.interceptor = (method, url, headers, [body]) {
         // checking sincce startWithContext makes a PUT and a GET request
         // to retrieve the current playbackstate
-        if (method == 'PUT') {
-          expect(method, 'PUT');
+        if (method == HttpMethod.PUT) {
+          expect(method, HttpMethod.PUT);
           expect(body, isNotNull);
           expect(body,
               '{"context_uri":"contextUri","offset":{"uri":"urioffset"}}');
@@ -503,8 +515,8 @@ Future main() async {
       spotify.interceptor = (method, url, headers, [body]) {
         // checking sincce startWithTracks makes a PUT and a GET request
         // to retrieve the current playbackstate
-        if (method == 'PUT') {
-          expect(method, 'PUT');
+        if (method == HttpMethod.PUT) {
+          expect(method, HttpMethod.PUT);
           expect(body, isNotNull);
           expect(body, '{"uris":["track1"],"position_ms":10}');
         }